+++ Block/Adminhtml/Form/Edit/Tab/Fields/Form.php
@@ -354,7 +354,7 @@ class Form extends \Alekseon\AlekseonEav\Block\Adminhtml\Entity\Edit\Form
             return true;
         }
 
-        return true;
+        return false;
     }
 
     /**