+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 ## [Unreleased]
+### Added
+- small alekseon logo
 
 ## [102.3.0] - 2023-03-03
 ### Added