+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - small alekseon logo
+- cache tags
 
 ## [102.3.0] - 2023-03-03
 ### Added